bitmapをモノクログリッドに変換する処理をクラス化した
